refactor: query key, option 세그먼트 이동
