+++ apps/web/src/entities/alumni-contacts/api/index.ts
@@ -1,3 +1,1 @@
 export { getAlumniContacts } from './get';
-export { alumniContactsQueryOptions } from './options';
-export { alumniContactsQueryKeys } from './keys';
