+++ apps/web/src/entities/alumni-contacts/config/query/alumniContactsKeys.ts
@@ -1,4 +1,4 @@
-import { GetAlumniContactsQuery } from '../types';
+import { GetAlumniContactsQuery } from '../../types';
 
 export const alumniContactsQueryKeys = {
   all: ['alumni-contacts'] as const,
