+++ apps/web/src/entities/alumni-contacts/config/query/index.ts
@@ -1,0 +1,2 @@
+export { alumniContactsQueryOptions } from './alumniContactsOptions';
+export { alumniContactsQueryKeys } from './alumniContactsKeys';
